tpfinal/requirements.txt
Django==3.2.9
pymysql
cryptography
djangorestframework
django-cors-headers
